feat(splash-sections): restructure and enhance splash sections for improved user experience
- Moved existing sections to a new directory structure under splash-sections for better organization.
- Introduced new components: AnalyticsSection, ChatDemoSection, CTASection, FeaturesSection, FooterSection, HeroSection, and PricingSection to enhance the home page layout.
- Updated imports in HomePageClient to reflect the new structure.
- Enhanced accessibility and responsiveness across all new sections, ensuring a cohesive user experience.
