+++ src/components/pages/home-page-client.tsx
@@ -4,13 +4,13 @@ import type React from "react";
 import { useState, useEffect } from "react";
 import { AnimatedBackground } from "@/components/ui/AnimatedBackground";
 import { PageLoader } from "@/components/ui/page-loader";
-import { HeroSection } from "@/components/sections/HeroSection";
-import { FeaturesSection } from "@/components/sections/FeaturesSection";
-import { PricingSection } from "@/components/sections/PricingSection";
-import { ChatDemoSection } from "@/components/sections/ChatDemoSection";
-import { AnalyticsSection } from "@/components/sections/AnalyticsSection";
-import { CTASection } from "@/components/sections/CTASection";
-import { FooterSection } from "@/components/sections/FooterSection";
+import { HeroSection } from "@/components/splash-sections/HeroSection";
+import { FeaturesSection } from "@/components/splash-sections/FeaturesSection";
+import { PricingSection } from "@/components/splash-sections/PricingSection";
+import { ChatDemoSection } from "@/components/splash-sections/ChatDemoSection";
+import { AnalyticsSection } from "@/components/splash-sections/AnalyticsSection";
+import { CTASection } from "@/components/splash-sections/CTASection";
+import { FooterSection } from "@/components/splash-sections/FooterSection";
 import { EmailSignupModal } from "@/components/modals/EmailSignupModal";
 import { MascotVibesOverlay } from "@/components/modals/MascotVibesOverlay";
 import BetaSignupModal from "@/components/BetaSignupModal";
