+++ src/components/splash-sections/AnalyticsSection.tsx
