+++ src/components/splash-sections/ChatDemoSection.tsx
