+++ src/components/splash-sections/FeaturesSection.tsx
