+++ src/components/splash-sections/FooterSection.tsx
