+++ src/components/splash-sections/HeroSection.tsx
