+++ src/components/splash-sections/PricingSection.tsx
